Create Rotate_Matrix_to_90_Degree
Program Performed - Matrix Rotation
language Used - Java
Program theory
First create a java program to rotate a matrix by 90 degree in a clockwise rotation
clockwise rotation is also known as right rotation

The rotation of a matrix involves two steps:
First, find the transpose of the given matrix.
Swap the elements of the first column with the last column (if the matrix is of 3*3). The second column remains the same.

Note: Matrix must have the same number of rows and columns.